+++ packages/components/CHANGELOG.md
@@ -1,3 +1,6 @@
+# unreleased
+- Pagination no longer considers `0` a valid input.
+
 # 2.0.0
 - Chart legend component now uses withInstanceId HOC so the ids used in several HTML elements are unique.
 - Chart component now accepts data with negative values.